This update shows dAPI chain as an inclusive MD file This used to use a Vue component but the component runs in the browser so the data is not available to the flex build. The inclusive MD file is part of the Vite build. The chains list is now pre-built using an axios script so the browser does not need to call the endpoint in operations.
