--- docs/reference/dapis/chains.md
@@ -1,26 +1,0 @@
----
-title: Chains and Contracts
-sidebarHeader: Reference
-sidebarSubHeader: dAPIs
-pageHeader: Reference → dAPIs
-path: /reference/dapis/chains.html
-outline: deep
-tags:
----
-
-<PageHeader/>
-
-<SearchHighlight/>
-
-<FlexStartTag/>
-
-# {{$frontmatter.title}}
-
-dAPIs can be read on the following chains using
-[API3 Market proxy contracts](/reference/dapis/understand/proxy-contracts.md).
-Use the [API3 Market<ExternalLinkImage/>](https://market.api3.org/) website to
-create proxy contracts.
-
-<ChainsList/>
-
-<FlexEndTag/>
