+++ docs/reference/dapis/sidebar.js
@@ -36,42 +36,8 @@ module.exports = [
 
       {
         text: 'Chains and Contracts',
-        link: '/reference/dapis/chains',
+        link: '/reference/dapis/chains/',
       },
     ],
   },
-  /*{
-    text: 'DapiServer Contract',
-    collapsed: false,
-    items: [
-      {
-        text: 'Overview',
-        link: '/reference/dapis/functions/',
-      },
-      {
-        text: '• Read Datafeed with DapiName',
-        link: '/reference/dapis/functions/read-data-feed-with-dapi-name',
-      },
-      {
-        text: '• Read Datafeed Value with dAPI Name',
-        link: '/reference/dapis/functions/read-data-feed-value-with-dapi-name',
-      },
-      {
-        text: '• Read Datafeed with Id',
-        link: '/reference/dapis/functions/read-data-feed-with-id',
-      },
-      {
-        text: '• Read Datafeed Value with Id',
-        link: '/reference/dapis/functions/read-data-feed-value-with-id',
-      },
-      {
-        text: '• Reader Can Read Datafeed',
-        link: '/reference/dapis/functions/reader-can-read-datafeed',
-      },
-      {
-        text: '• Datafeed Id to Reader to Whitelist Status',
-        link: '/reference/dapis/functions/data-feed-id-to-reader-to-whitelist-status',
-      },
-    ],
-  },*/
 ];
